Moved BottomDots to components folder, green color
